refactor: 도메인과 뷰 분리하기
